Bring the remove site back
